bug(appfix) -re-order document table -abstract sweetalert object -remove parameter abbreviations -flesh out functionalities in readme -remove api key warning from tinymce -empty search -search not fund notification -eslint fix -more tests -test language correction
