+++ client/public/index.html
@@ -11,7 +11,7 @@
        
       <title>Doc-Manager</title>
       <!--Let browser know website is optimized for mobile-->
-  <script src="https://cloud.tinymce.com/stable/tinymce.min.js"></script>
+  <script src="https://cloud.tinymce.com/stable/tinymce.min.js?apiKey=2sy53b69f34eq5k94nyfkv0n43td0qh4si39wzrfyk7qbq3j"></script>
   <script>tinymce.init({ selector:'textarea' });</script>
       <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     </head>
