+++ client/tests/component/footer.spec.js
@@ -5,16 +5,16 @@ import { shallow } from 'enzyme';
 import Footer from '../../src/components/layout/Footer.jsx';
 /* eslint-disable no-unused-expressions*/
 
-describe('Footer component with the elements available', () => {
+describe('Footer component', () => {
   const wrapper = shallow(<Footer />);
-  it('it should render div elements', () => {
+  it('should render div elements', () => {
     expect(wrapper.find('div')).to.exist;
   });
 
-  it('it should contain a footer tag', () => {
+  it('should contain a footer tag', () => {
     expect(wrapper.find('footer')).to.exist;
   });
-    it('it should contain a link tag', () => {
+    it('should contain a link tag', () => {
     expect(wrapper.find('a')).to.exist;
   });
 });
