+++ client/tests/component/manageDocuments.spec.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import sinon from 'sinon'; 
-import ManageDoc from '../../src/components/sections/ManageDoc.jsx';
+import ManageDocuments from '../../src/components/sections/ManageDocuments.jsx';
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
-describe('<ManageDoc />', () => {
+describe('<ManageDocuments />', () => {
   let props;
   let store;
   let copyStore;
@@ -28,35 +28,35 @@ describe('<ManageDoc />', () => {
     };
      wrapper = mount(
       <Provider store={store} >
-        <ManageDoc {...props} />
+        <ManageDocuments {...props} />
       </Provider>
     );
   });
   it('should connect to the redux store', () => {
     expect(wrapper.props().store.getState()).to.eql(copyStore);
   });
-  it('it should render div elements', () => {
+  it('should render div elements', () => {
      expect(wrapper.find('div')).to.exist;
     });
-  it('it should render img element', () => {
+  it('should render img element', () => {
      expect(wrapper.find('img')).to.exist;
     });
-  it('it should render Paginate element', () => {
+  it('should render Paginate element', () => {
      expect(wrapper.find('Paginate')).to.exist;
     });
 
-  sinon.spy(ManageDoc.prototype, 'componentWillReceiveProps');
-  sinon.spy(ManageDoc.prototype, 'render');
+  sinon.spy(ManageDocuments.prototype, 'componentWillReceiveProps');
+  sinon.spy(ManageDocuments.prototype, 'render');
 
-  it(' componentWillReceiveProps exists', () => {
-    expect(ManageDoc.prototype.componentWillReceiveProps.calledOnce).to.exist;
+  it(' componentWillReceiveProps function exists', () => {
+    expect(ManageDocuments.prototype.componentWillReceiveProps.calledOnce).to.exist;
   });
    it(' Render exists', () => {
-    expect(ManageDoc.prototype.render.calledOnce).to.exist;
+    expect(ManageDocuments.prototype.render.calledOnce).to.exist;
   });
-  it('componentWillUnmount exists', () => {
+  it('ComponentWillUnmount function exists', () => {
     wrapper.unmount();
-    sinon.spy(ManageDoc.prototype, 'componentWillUnmount');
-    expect(ManageDoc.prototype.componentWillUnmount.calledOnce);
+    sinon.spy(ManageDocuments.prototype, 'componentWillUnmount');
+    expect(ManageDocuments.prototype.componentWillUnmount.calledOnce);
   });
 });
