+++ reports/CHROME_59.0.3071.115_MAC_admin.xml
@@ -5,7 +5,7 @@
 
   <testsuite name="admin"
     errors="0" failures="0" hostname="" id="" package="admin" skipped="0"
-    tests="0" time="0.000" timestamp="Tue, 18 Jul 2017 13:18:04 GMT">
+    tests="0" time="0.000" timestamp="Tue, 18 Jul 2017 15:25:22 GMT">
