+++ reports/CHROME_59.0.3071.115_MAC_user.xml
@@ -5,9 +5,9 @@
 
   <testsuite name="user"
     errors="0" failures="0" hostname="" id="" package="user" skipped="0"
-    tests="7" time="32.30" timestamp="Tue, 18 Jul 2017 13:18:05 GMT">
+    tests="7" time="31.51" timestamp="Tue, 18 Jul 2017 15:25:23 GMT">
   
-    <testcase name="User sign up without credentials" classname="user" time="6.466" assertions="4">
+    <testcase name="User sign up without credentials" classname="user" time="6.112" assertions="4">
 
     
 
@@ -18,7 +18,7 @@
     
     </testcase>
   
-    <testcase name="User sign up with wrong email" classname="user" time="3.211" assertions="3">
+    <testcase name="User sign up with wrong email" classname="user" time="3.040" assertions="3">
 
     
 
@@ -27,7 +27,7 @@
     
     </testcase>
   
-    <testcase name="User sign up with wrong password" classname="user" time="3.113" assertions="3">
+    <testcase name="User sign up with wrong password" classname="user" time="2.881" assertions="3">
 
     
 
@@ -36,36 +36,7 @@
     
     </testcase>
   
-    <testcase name="Successful sign up" classname="user" time="3.696" assertions="4">
-
-    
-
-    
-
-    
-
-    
-    </testcase>
-  
-    <testcase name="User should be able to update his profile" classname="user" time="5.522" assertions="13">
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
-
-    
+    <testcase name="Successful sign up" classname="user" time="3.560" assertions="4">
 
     
 
@@ -76,7 +47,13 @@
     
     </testcase>
   
-    <testcase name="User should be able to create new document" classname="user" time="7.272" assertions="10">
+    <testcase name="User should be able to update his profile" classname="user" time="5.468" assertions="13">
+
+    
+
+    
+
+    
 
     
 
@@ -99,7 +76,30 @@
     
     </testcase>
   
-    <testcase name="User sign out of application" classname="user" time="3.018" assertions="8">
+    <testcase name="User should be able to create new document" classname="user" time="7.026" assertions="10">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    </testcase>
+  
+    <testcase name="User sign out of application" classname="user" time="3.427" assertions="8">
