+++ server/tests/controllers/document.spec.js
@@ -1,4 +1,7 @@
+/* global Documents:true */
+
 const app = require('../../../app');
+
 const expect = chai.expect;
 const api = request(app);
 
@@ -8,14 +11,14 @@ const Admin = generateAdminRole();
 const Regular = generateRegularRole();
 const Staff = generateStaffRole();
 const document = generateDocument(user2data.id);
-const userDocument =  generateDocument(user1data.id);
-const privateDocument =  generateDocument(user1data.id);
+const userDocument = generateDocument(user1data.id);
+const privateDocument = generateDocument(user1data.id);
 
 describe('Document controller', () => {
   const user = {};
   const adminUser = {};
   beforeEach((done) => {
-    sequelize.sync({ force: true})
+    sequelize.sync({ force: true })
       .done(() => {
         Role.bulkCreate([{ title: 'admin' }, { title: 'regular' }])
           .then(() => {
@@ -28,34 +31,31 @@ describe('Document controller', () => {
                   .then((createdAdmin) => {
                     adminUser.id = createdAdmin.id;
                     adminUser.token = generateToken(createdAdmin);
-                    document.userId = createdAdmin.id
+                    document.userId = createdAdmin.id;
                     Documents.create(document)
-                    .then((createdDoc) => {
+                      .then((createdDoc) => {
                         document.id = createdDoc.id;
                         adminUser.id = createdDoc.userId;
                         userDocument.userId = user.id;
                         Documents.create(userDocument)
                           .then((createdDoc) => {
-                              userDocument.id = createdDoc.id;
-                              user.id = createdDoc.userId;
-                              privateDocument.userId = adminUser.id;
-                              privateDocument.access = 'Private';
-                         Documents.create(privateDocument)
-                          .then((createdDoc) => {
-                              privateDocument.id = createdDoc.id;
-                              done();
+                            userDocument.id = createdDoc.id;
+                            user.id = createdDoc.userId;
+                            privateDocument.userId = adminUser.id;
+                            privateDocument.access = 'Private';
+                            Documents.create(privateDocument)
+                              .then((createdDoc) => {
+                                privateDocument.id = createdDoc.id;
+                                done();
+                              });
                           });
-                          })
-                })
-                
-
-                   
+                      });
                   });
               });
           });
       });
   });
-   afterEach((done) => {
+  afterEach((done) => {
     User.destroy({ where: {} })
       .then(() => {
         Role.destroy({ where: {} })
@@ -65,134 +65,131 @@ describe('Document controller', () => {
       });
   });
 
-describe(' create document', () => {
-  it('should create documents',(done) => {
-    api.post('/api/v1/documents')
-    .send(document)
-      .set({authorization : adminUser.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(201);
-      done();
-      }); 
-  });
-  it('should not create documents without title',(done) => {
-    const newDocument = {...document};
-    delete newDocument.title;
-    api.post('/api/v1/documents')
-    .send(newDocument)
-      .set({authorization : user.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(400);
-      done();
-      }); 
-  });
-  it('should not create documents without access', (done) => {
-    const newDocument = {...document};
-    delete newDocument.access;
-    api.post('/api/v1/documents')
-      .send(newDocument)
-        .set({authorization : user.token})
+  describe(' Document', () => {
+    it('should create documents', (done) => {
+      api.post('/api/v1/documents')
+        .send(document)
+        .set({ authorization: adminUser.token })
         .end((err, res) => {
-        expect(res.status).to.equal(400);
-        done();
+          expect(res.status).to.equal(201);
+          done();
+        });
+    });
+    it('should not create documents without title', (done) => {
+      const newDocument = { ...document };
+      delete newDocument.title;
+      api.post('/api/v1/documents')
+        .send(newDocument)
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(400);
+          done();
+        });
+    });
+    it('should not create documents without access', (done) => {
+      const newDocument = { ...document };
+      delete newDocument.access;
+      api.post('/api/v1/documents')
+        .send(newDocument)
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(400);
+          done();
         });
     });
     it('should not create documents with a wrong access option', (done) => {
-      document.access = "Hi"
-    api.post('/api/v1/documents')
-      .send(document)
-        .set({authorization : user.token})
+      document.access = 'Hi';
+      api.post('/api/v1/documents')
+        .send(document)
+        .set({ authorization: user.token })
         .end((err, res) => {
-        expect(res.status).to.equal(400);
-        done();
+          expect(res.status).to.equal(400);
+          done();
         });
     });
-})
+  });
 
-describe('find document', (done) => {
-  it('should get documents',(done) => {
-    api.get(`/api/v1/documents/${document.id}`)
-      .set({authorization : user.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(200);
-      done();
-    }); 
-  });
-  it('should throw a not found error for documents that do not exist',(done) => {
-    api.get('/api/v1/documents/09099')
-      .set({authorization : user.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(404);
-      done();
-      }); 
-  });
-  it('should not get document if its not owned by the user',(done) => {
-   const privateDoc = {...privateDocument, access :'Private'}
-    api.get(`/api/v1/documents/${privateDoc.id}`)
-      .set({authorization : user.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(401);
-      done();
-    }); 
+  describe('find document', () => {
+    it('should get documents with a specific id', (done) => {
+      api.get(`/api/v1/documents/${document.id}`)
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(200);
+          done();
+        });
+    });
+    it('should throw a not found error for documents that do not exist', (done) => {
+      api.get('/api/v1/documents/09099')
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(404);
+          done();
+        });
+    });
+    it('should not get document if its not owned by the user', (done) => {
+      const privateDoc = { ...privateDocument, access: 'Private' };
+      api.get(`/api/v1/documents/${privateDoc.id}`)
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(401);
+          done();
+        });
     });
   });
-  describe('delete document', (done) => {
-    it('should delete a document',(done) => {
+  describe('delete document', () => {
+    it('should delete a document', (done) => {
       api.delete(`/api/v1/documents/${userDocument.id}`)
-        .set({authorization : user.token})
+        .set({ authorization: user.token })
         .end((err, res) => {
           expect(res.status).to.equal(200);
           done(err);
-      });
+        });
     });
-    it('should not  delete a document',(done) => {
+    it('should not  delete a document if user is not authorized', (done) => {
       api.delete(`/api/v1/documents/${privateDocument.id}`)
-        .set({authorization : user.token})
+        .set({ authorization: user.token })
         .end((err, res) => {
-        expect(res.status).to.equal(401);
-        done();
-        }); 
+          expect(res.status).to.equal(401);
+          done();
+        });
     });
-    it('should throw a not found error for documents that do not exist',(done) => {
-     api.delete('/api/v1/documents/09099')
-      .set({authorization : user.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(404);
-      done();
-      }); 
+    it('should throw a not found error for documents that do not exist', (done) => {
+      api.delete('/api/v1/documents/09099')
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(404);
+          done();
+        });
+    });
   });
-});
-describe('update document', (done) => {
-    it('should update a document',(done) => {
+  describe('update document', () => {
+    it('should update a document', (done) => {
       api.put(`/api/v1/documents/${userDocument.id}`)
-      .send({title: 'hello'})
-        .set({authorization : user.token})
+        .send({ title: 'hello' })
+        .set({ authorization: user.token })
         .end((err, res) => {
           expect(res.status).to.equal(200);
           done(err);
-      });
+        });
     });
-    it('should not  update a document',(done) => {
+    it('should not update a document if not owned by the user', (done) => {
       api.put(`/api/v1/documents/${privateDocument.id}`)
-        .send({title: 'hello'})
-        .set({authorization : user.token})
+        .send({ title: 'hello' })
+        .set({ authorization: user.token })
         .end((err, res) => {
-        expect(res.status).to.equal(401);
-        done();
-        }); 
+          expect(res.status).to.equal(401);
+          done();
+        });
     });
-    it('should throw a not found error for documents that do not exist',(done) => {
-     api.put('/api/v1/documents/09099')
-      .send({title: 'hello'})
-      .set({authorization : user.token})
-      .end((err, res) => {
-      expect(res.status).to.equal(404);
-      done();
-      }); 
+    it('should throw a not found error for documents that do not exist', (done) => {
+      api.put('/api/v1/documents/09099')
+        .send({ title: 'hello' })
+        .set({ authorization: user.token })
+        .end((err, res) => {
+          expect(res.status).to.equal(404);
+          done();
+        });
+    });
   });
 });
-});
 
-
-
-
