+++ server/tests/models/role.spec.js
@@ -1,17 +1,17 @@
 describe('Roles Model', () => {
   afterEach((done) => {
     Role.destroy({
-        where: {}
+      where: {}
     })
-    .then(() => {
-      done();
-    });
+      .then(() => {
+        done();
+      });
   });
   it('should create a new roles', (done) => {
     Role.create({ title: 'admin' })
       .then((createdRole) => {
         expect(createdRole.title).to.eql('admin');
-        done();       
+        done();
       });
   });
 });
